行走书店v1.5

+++ sc2019/sc-provider/sc-xzsd-app/src/main/java/com/xzsd/app/goodsMessage/entity/GoodsMessage.java
@@ -65,7 +65,7 @@ public class GoodsMessage {
     /**
      * 商品评价星级 01差评 23中评 45好评
      */
-    private int goodsStar;
+    private float goodsStar;
     /**
      * 上架时间
      */
@@ -257,11 +257,11 @@ public class GoodsMessage {
         this.goodsView = goodsView;
     }
 
-    public int getGoodsStar() {
+    public float getGoodsStar() {
         return goodsStar;
     }
 
-    public void setGoodsStar(int goodsStar) {
+    public void setGoodsStar(float goodsStar) {
         this.goodsStar = goodsStar;
     }
